pyproject.toml
[project]
name = "localscript"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "chromadb>=1.5.5",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "langchain>=1.2.14",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.48",
]
